requirements.txt
Django==2.0.8
djangorestframework==3.9.0
drf_writable_nested==0.5.1
requests==2.20